Extract factor loadings from galamm object

Extract factor loadings from galamm object

Usage

# S3 method for class 'galamm'
factor_loadings(object)

Arguments

object

Object of class galamm returned from galamm.

Value

A matrix containing the estimated factor loadings with corresponding standard deviations.

Details

This function has been named factor_loadings rather than just loadings to avoid conflict with stats::loadings.

See also

Author

The example for this function comes from PLmixed, with authors Nicholas Rockwood and Minjeong Jeon (Rockwood and Jeon 2019) .

Examples

# Logistic mixed model with factor loadings, example from PLmixed
data("IRTsim", package = "PLmixed")

# Reduce data size for the example to run faster
IRTsub <- IRTsim[IRTsim$item < 4, ]
IRTsub <- IRTsub[sample(nrow(IRTsub), 300), ]
IRTsub$item <- factor(IRTsub$item)

# Fix loading for first item to 1, and estimate the two others freely
loading_matrix <- matrix(c(1, NA, NA), ncol = 1)

# Estimate model
mod <- galamm(y ~ item + (0 + ability | sid) + (0 + ability | school),
  data = IRTsub, family = binomial, load_var = "item",
  factor = "ability", lambda = loading_matrix
)

# Show estimated factor loadings, with standard errors
factor_loadings(mod)
#>           ability        SE
#> lambda1 1.0000000        NA
#> lambda2 0.3570000 0.2113691
#> lambda3 0.3713262 0.2206723