Fit a generalized additive latent and mixed model

This function fits a generalized additive latent and mixed model (GALAMMs), as described in Sørensen et al. (2023) . The building blocks of these models are generalized additive mixed models (GAMMs) (Wood 2017) , of which generalized linear mixed models (Breslow and Clayton 1993; Harville 1977; Henderson 1975; Laird and Ware 1982) are special cases. GALAMMs extend upon GAMMs by allowing factor structures, as commonly used to model hypothesized latent traits underlying observed measurements. In this sense, GALAMMs are an extension of generalized linear latent and mixed models (GLLAMMs) (Skrondal and Rabe-Hesketh 2004; Rabe-Hesketh et al. 2004) which allows semiparametric estimation. The implemented algorithm used to compute model estimates is described in Sørensen et al. (2023) , and is an extension of the algorithm used for fitting generalized linear mixed models by the lme4 package (Bates et al. 2015) . The syntax used to define factor structures is based on that used by the PLmixed package, which is detailed in Rockwood and Jeon (2019) .

Usage

galamm(
  formula,
  weights = NULL,
  data,
  family = gaussian,
  family_mapping = rep(1L, nrow(data)),
  load.var = NULL,
  lambda = NULL,
  factor = NULL,
  start = NULL,
  control = galamm_control()
)

Arguments

formula

A formula specifying the model. Smooth terms are defined in the style of the mgcv and gamm4 packages, see (Wood 2017) for an introduction. Random effects are specifified using lme4 syntax, which is described in detail in (Bates et al. 2015) . Factor loadings will also be part of the model formula, and is based on the syntax of the PLmixed package (Rockwood and Jeon 2019) .

weights

An optional formula object specifying an expression for the residual variance. Defaults to NULL, corresponding to homoscedastic errors. The formula is defined in lme4 style; see vignettes and examples for details.

data

A dataset containing all the variables specified by the model formula, with the exception of factor loadings.

family

A vector containing one or more model families. For each element in family there should be a corresponding element in family_mapping specifying which elements of the response are conditionally distributed according to the given family. Currently family can be one of gaussian, binomial, and poisson, and only canonical link functions are supported.

family_mapping

A vector mapping from the elements of family to rows of data. Defaults to rep(1L, nrow(data)), which means that all observations are distributed according to the first element of family.

load.var

Character specifying the name of the variable in data identifying what the factors load onto. That is, each unique value of load.var corresponds to a unique factor loading. Currently only a single loading

lambda

List of factor loading matrices. Numerical values indicate that the given value is fixed, while NA means that the entry is a parameter to be estimated.

factor

List of character vectors identical to the factor loadings specified in formula. For each list element, the \(j\)th entry in the character vector corresponds to the \(j\)th column of the corresponding matrix in lambda.

start

A named list of starting values for parameters. Possible names of list elements are "theta", "beta", "lambda", and "weights", all of should be numerical vectors with starting values. Default to NULL, which means that some relatively sensible defaults are used.

control

Control object for the optimization procedure of class galamm_control resulting from calling galamm_control.

Value

A model object of class galamm.

References

Bates DM, Mächler M, Bolker B, Walker S (2015). “Fitting Linear Mixed-Effects Models Using Lme4.” Journal of Statistical Software, 67(1), 1--48. ISSN 1548-7660, doi:10.18637/jss.v067.i01 , 2020-11-25.

Breslow NE, Clayton DG (1993). “Approximate Inference in Generalized Linear Mixed Models.” Journal of the American Statistical Association, 88(421), 9--25. ISSN 0162-1459, doi:10.2307/2290687 , 2290687, 2020-12-02.

Harville DA (1977). “Maximum Likelihood Approaches to Variance Component Estimation and to Related Problems.” Journal of the American Statistical Association, 72(358), 320--338. ISSN 0162-1459, doi:10.2307/2286796 , 2286796, 2022-07-22.

Henderson CR (1975). “Best Linear Unbiased Estimation and Prediction under a Selection Model.” Biometrics, 31(2), 423--447. ISSN 0006-341X, doi:10.2307/2529430 , 2529430, 2022-07-22.

Laird NM, Ware JH (1982). “Random-Effects Models for Longitudinal Data.” Biometrics, 38(4), 963--974. ISSN 0006-341X, doi:10.2307/2529876 , 2529876, 2020-11-25.

Rabe-Hesketh S, Skrondal A, Pickles A (2004). “Generalized Multilevel Structural Equation Modeling.” Psychometrika, 69(2), 167--190. ISSN 1860-0980, doi:10.1007/BF02295939 , 2020-11-03.

Rockwood NJ, Jeon M (2019). “Estimating Complex Measurement and Growth Models Using the R Package PLmixed.” Multivariate Behavioral Research, 54(2), 288--306. ISSN 0027-3171, doi:10.1080/00273171.2018.1516541 , 2020-11-25.

Skrondal A, Rabe-Hesketh S (2004). Generalized Latent Variable Modeling, Interdisciplinary Statistics Series. Chapman and Hall/CRC, Boca Raton, Florida.

Sørensen Ø, Fjell AM, Walhovd KB (2023). “Longitudinal Modeling of Age-Dependent Latent Traits with Generalized Additive Latent and Mixed Models.” Psychometrika, 88(2), 456--486. ISSN 1860-0980, doi:10.1007/s11336-023-09910-z , 2023-05-25.

Wood SN (2017). Generalized Additive Models: An Introduction with R, 2 edition. Chapman and Hall/CRC.