GALAMM plotting

This function plots fitted galamm objects. It uses mgcv::plot.gam, but the confidence bands take uncertainty about factor loadings into account.

Usage

# S3 method for galamm
plot(x, ...)

Arguments

x

A fitted galamm object as produced by galamm.

...

Other arguments passed on to mgcv::plot.gam.

Value

A plot is generated.

Examples

# Example dataset with three measurements of a latent response
# Define loading matrix
load.mat <- matrix(c(1, NA, NA), ncol = 1)
dimnames(load.mat) <- list(c("item1", "item2", "item3"), NULL)
if (FALSE) {
mod <- galamm(
  formula = y ~ s(x, by = weight),
  random = ~(1|id),
  data = dat1,
  load_var = "item",
  lambda = load.mat,
  factor = "weight",
  lambda_init = c(2, .4),
  optim_control = list(trace = 3)
  )
plot(mod)
  }