Fit a generalized additive latent and mixed model

Fits a GALAMM using a profile likelihood algorithm. gamm4::gamm4 is used to fit the underling generalized additive mixed model at each step.

Usage

galamm(
  formula,
  random = NULL,
  family = stats::gaussian(),
  data = list(),
  load_var = NULL,
  lambda = NULL,
  factor = NULL,
  lambda_init = NULL,
  optim_method = "L-BFGS-B",
  optim_control = list(),
  numDeriv_method = "simple"
)

Arguments

formula

A formula for the fixed effects, including smooth terms.

random

A formula for random effects.

family

A GLM family.

data

A data frame containing the variables used in the model.

load_var

The variables which the factors load onto.

lambda

A factor loading matrix, with row names corresponding to load_var.

factor

The latent variable, or factor.

lambda_init

Initial value for free factor loadings. Default to NULL, which means that they are randomly drawn from a uniform distribution with support \([0,1]\).

optim_method

Passed on to stats::optim.

optim_control

Passed on to stats::optim.

numDeriv_method

Passed on to numDeriv::jacobian.

Value

An object of S3 class galamm.

Details

See the documentation to gamm4::gamm4 and PLmixed::PLmixed for details about the syntax.

Examples

# Example dataset with three measurements of a latent response
# Define loading matrix
load.mat <- matrix(c(1, NA, NA), ncol = 1)
dimnames(load.mat) <- list(c("item1", "item2", "item3"), NULL)
if (FALSE) {
mod <- galamm(
  formula = y ~ s(x, by = weight),
  random = ~(1|id),
  data = dat1,
  load_var = "item",
  lambda = load.mat,
  factor = "weight",
  lambda_init = c(2, .4),
  optim_control = list(trace = 3)
  )
  }